Real Tracking implemented
Tracking is now much better, but is lost in some cases (like too when
the image is not bright enough. Tracking is not updated if new found
Pupil is too far away from last tracked pupil. This is not optimal at
all, but better then before. The biggest problem seems to be the
brightness in the image and other dark areas (eyebrow for instance).
Tracking is still very shaky. It seems that the pupil cannot be
segmented properly. IR could improve this drastically...