Ha! Got much more stable tracking now
Basic signal processing did the trick. I'm now computing the mean
of the current position with the old position. This makes the
tracking a bit slower, but also a lot smoother. With additional
restricting the movement (movement must be above a certain
threshold) its pretty smooth and not that bad at all. I'm proud of
myself.